+++ plydata/dataframe/__init__.py
@@ -3,4 +3,5 @@ Verb implementations for the dataframe
 """
 from .one_table import *     # noqa
 from .two_table import *     # noqa
+from .tidy import *          # noqa
 from .helpers import *       # noqa